[DEV] Reconstruct project
