+++ src/utils/model.py
@@ -1,7 +1,5 @@
 import os
 import random
-import logging
-import datetime
 import numpy as np
 from scipy.sparse import csr_matrix
 from scipy.sparse.linalg import eigs
@@ -11,6 +9,7 @@ import torch
 import torch.nn.functional as F
 import torch.nn as nn
 from torch_geometric.nn import SAGEConv
+
 
 def seed_everything(seed=0):
     random.seed(seed)
@@ -22,19 +21,6 @@ def seed_everything(seed=0):
     torch.backends.cudnn.deterministic = True
     torch.backends.cudnn.benchmark = False
 
-def setup_logger(name, log_dir='logs'):
-    os.makedirs(log_dir, exist_ok=True)
-    log_file = os.path.join(log_dir, '{}_{}.log'.format(datetime.datetime.now().strftime('%m%d%H%M'), name))
-    
-    logging.basicConfig(
-        level=logging.INFO,
-        format='%(asctime)s - %(levelname)s - %(message)s',
-        handlers=[
-            logging.FileHandler(log_file, mode='w'),
-            logging.StreamHandler()
-        ]
-    )
-    return logging.getLogger(name)
 
 class Net(nn.Module):
     def __init__(self):
@@ -53,6 +39,7 @@ class Net(nn.Module):
     def forward(self, input):
         return self.main(input)
 
+
 class MLP(nn.Module):
     def __init__(self, in_channels, out_channels, mid_channels=128):
         super(MLP, self).__init__()
@@ -65,6 +52,7 @@ class MLP(nn.Module):
     def forward(self, input):
         x = self.main(input)
         return x
+
 
 class GraphSAGE(torch.nn.Module):
     def __init__(self, m):
@@ -92,8 +80,7 @@ class GraphSAGE(torch.nn.Module):
         x = self.conv3(x, edge_index)
         return x
 
-# no1 方案十九：简单缩放 (SimpleScale)
-# 仅使用密度进行缩放 (Scaling)，不进行平移 (Shift)，简化版 AdaIN
+
 class GraphSAGE_SimpleScale(torch.nn.Module):
     def __init__(self, m):
         super(GraphSAGE_SimpleScale, self).__init__()
@@ -110,8 +97,7 @@ class GraphSAGE_SimpleScale(torch.nn.Module):
         self.conv3 = SAGEConv(128, 64)
         self.conv3.lin_l.weight.data.normal_(0, 1e-3)
         self.conv3.lin_r.weight.data.normal_(0, 1e-3)
-        
-        # 增加输出MLP层以增强表达能力
+
         self.output_mlp = nn.Sequential(
             nn.Linear(64, 32),
             nn.LeakyReLU(),
@@ -122,23 +108,24 @@ class GraphSAGE_SimpleScale(torch.nn.Module):
         x, edge_index = data.x, data.edge_index
         features = x[:, :-3]
         density = x[:, -3:]
-        
+
         x = self.conv1(features, edge_index)
         s1 = self.scale1(density)
         x = x * (1 + s1)
         x = F.leaky_relu(x)
         x = F.dropout(x, p=0.3, training=self.training)
-        
+
         x = self.conv2(x, edge_index)
         s2 = self.scale2(density)
         x = x * (1 + s2)
         x = F.leaky_relu(x)
         x = F.dropout(x, p=0.3, training=self.training)
-        
+
         x = self.conv3(x, edge_index)
         x = F.leaky_relu(x)
         x = self.output_mlp(x)
         return x
+
 
 def moon(n):
     m = int(n / 2)
@@ -152,12 +139,14 @@ def moon(n):
     n = len(x)
     return x, n
 
+
 def stationary(A):
     eig = eigs(A.T)
     ind = eig[0].real.argsort()[-1]
     est = eig[1][:, ind].real
     pr = est / est.sum() * A.shape[0]
     return pr
+
 
 def reconstruct(K, pr, n, m, fr, to):
     selected = np.random.choice(np.arange(n), m, replace=False)
@@ -179,6 +168,7 @@ def reconstruct(K, pr, n, m, fr, to):
     rec_orig[selected] = rec_unnormalized
     rec_orig[unselected] = rec_unnormalized[spd[:, unselected].argmin(0)]
     return rec_orig
+
 
 def dG(A, B):
     S = A.T @ B
